poky-lsb.conf: Instead of += with _append
The default variable DISTRO_FEATHURE was set in poky.conf before \
moving to defaultsetup.conf. Defaultsetup.conf is included after poky-lsb.conf \
So poky-lsb.conf sets the variable using += but since its empty, \
it might as well be an = and then the later ?= doesn't do anything since the \
variable is already set.

Signed-off-by: Xiaofeng Yan <xiaofeng.yan@windriver.com>