busybox: Introduce poky-tiny defconfig override
When building very small systems, it can be useful to spawn a shell from
a simple init script, rather than a full System V Init process. This
requires the shell be the session leader and be able to open the
controlling terminal if it is to have job control.

Create a busybox bbappend in meta-yocto with a poky-tiny/defconfig.
This adds SETSID and CTTYHACK for poky-tiny on top of the oe-core
defconfig.

Signed-off-by: Darren Hart <dvhart@linux.intel.com>
CC: Tim Bird <tim.bird@am.sony.com>
CC: Thomas Frydrych <tf+lists.yocto@r-finger.com>
CC: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
CC: Paul Eggleton <paul.eggleton@linux.intel.com>
CC: Phil Blundell <philb@gnu.org>
CC: Khem Raj <raj.khem@gmail.com>
CC: Koen Kooi <koen@dominion.thruhere.net>