atom-pc: Drop glibc --with-tls option, its now the only option for glibc
This option is unused by (e)glibc since 2011 and is the default. It has been
shown to interact badly with the configure option in ia32-base.inc from meta-intel
causing a rebuild of the whole system despite the only change being an
assignment with += vs =. The easiest fix is simply to drop it.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>