README: Clarify where to send patches
There is still a great deal of confusion over where patches for the
various components of the poky repository should be sent. Attempt to
mitigate this by including a blurb in the poky README similar to what
other layers and repositories do.

V2: Call out the special care required when dealing with the scripts
    directory per Saul Wold's recommendation.

V3: Recommend applying patches to oe-core prior to sending patches to
    the oe-core mailing list.

Signed-off-by: Darren Hart <dvhart@linux.intel.com>
CC: Richard Purdie <richard.purdie@intel.com>
CC: Paul Gortmaker <paul.gortmaker@windriver.com>
CC: Paul Eggleton <paul.eggleton@linux.intel.com>
CC: Khem Raj <raj.khem@gmail.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>