poky.conf: Remove git proxy configuration check url
The git proxy check seems to be hurting the user experience more than
its helping it so lets remove it andonly check http/https. Most builds
should be able to work from the http:// urls from the mirrors.

This also brings some parity to the situation as svn:// were not being
checked.

[YOCTO #3119]

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>