meta-yocto/linux-yocto: update qemu BSPs to 3.8 kernel
Change linux-yocto preferred version to 3.8 for the qemu BSPs. The
non-qemu default stays at 3.4 until the hardware reference platforms
can be validated on the 3.8 kernel.

At the same time, we remove the 3.0 linux-yocto bbappends, since they
have been removed in the core layers and will cause dangling append
problems if left in place.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>
Signed-off-by: Saul Wold <sgw@linux.intel.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>