poky.conf: fix WARN_QA and ERROR_QA settings
In OE-Core master revision 8f5675e6d3eb the bug that tests we did not
mention in ERROR_QA were being shown as warnings even if they weren't in
WARN_QA was fixed; however a number of warnings we want to be enabled
weren't in Poky's WARN_QA and thus were not shown after that change;
additionally some of these were set in the default value of ERROR_QA in
insane.bbclass (i.e. OE-Core's default) but would have previously shown
up as warnings in Poky because they weren't in ERROR_QA there.

To fix this, put the appropriate values back into WARN_QA / ERROR_QA
with a lean towards ERROR_QA - we want to be strict so that we keep
OE-Core clean of QA issues during recipe maintenance.

Fixes [YOCTO #4752].

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>