poky.conf: Don't force the addition of extra DISTRO_FEATURES
The DISTRO_FEATURES shouldn't unconditionally append items to it. This
makes it impossible to override it inside of local.conf or in a distro
based on Poky.

This moved the definition to poky.conf and created a new variable
(POKY_DEFAULT_DISTRO_FEATURES) which easy overriding of this, for Poky
based distros and used the 'DISTRO_FEATURES_DEFAULT' variable to avoid
duplication OE-Core default.

This makes the override of default DISTRO_FEATURES easier. User can now do:

,----[ Usage example for local.conf ]
| DISTRO_FEATURES = "${DISTRO_FEATURES_DEFAULT} ${DISTRO_FEATURES_LIBC} \
|                    largefile opengl multiarch"
`----

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>