oeqa/controllers/beaglebonetarget.py: fix deploy when kernel file is a broken link
Sometimes the uncompressed rootfs image contains boot/uImage but if it's a symlink, it seems broken from the master image's point of view.
Because [ -e /mnt/testrootfs/boot/uImage ] will return 1, it will proceed to copy the backup kernel file over it but cp ~/test-kernel /mnt/testrootfs/boot/uImage will also fail with:
cp: not writing through dangling symlink

Adding a new check here, if the boot/uImage is a link, leave it as is.

Signed-off-by: Corneliu Stoicescu <corneliux.stoicescu@intel.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>