poky.conf: resolve mismatch between poky bblayers version and OE-Core
Drop our custom LAYER_CONF_VERSION (since the OE-Core version is now the
same) and ensure that poky-sanity's update function runs first so that
anyone still using version 5 with poky gets the meta-yocto /
meta-yocto-bsp split handled instead of OE-Core's no-op upgrade. Also
fix the version check so that the poky 5->6 upgrade is still performed
if the latest version is greater than 6.

NOTE: this of course relies on the corresponding version bump patch in
OE-Core.

Fixes [YOCTO #6139].

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>