u-boot/beaglebone: Select Beaglebone Black target specifically.
The am335x_evm_config target has a subtle problem which is not present
in the am335x_boneblack_config target: When booting from an external
MMC card, the internal MMC is not accessible using the former target.

This problem was introduced in U-Boot in commit 80b24fcd3083515e6b961,
due to the addition of the CONFIG_DM_MMC option, and the
am335x_boneblack_config target does not have this problem.

Signed-off-by: Kristian Amlie <kristian.amlie@mender.io>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>