layer.conf: Add the 'wrl' tag in the LAYERSERIES_COMPAT
Issue: US117235

Even this is not a layer owned by Wind River, but we do claim that the
distros and machines in this layer are supported by Wind River. So
add the 'wrl' tag in the LAYERSERIES_COMPAT to make sure that the
distros and machines in this layer are listed by default in the
output of '--list-{machines,distros}'.

(LOCAL REV: NOT UPSTREAM) -- Wind River Specific

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>