meta-yocto/conf/layer.conf: define a dummy collection
Issue: LIN9-7620

Fixed:
./update.py -b WRLINUX_9_BASE_RCPL0019 -l meta-yocto
ERROR: Unable to find BBFILE_COLLECTIONS value in initial output

The meta-yocto layer is in the sub dir of meta-yocto repo, which is
meta-yocto/meta-yocto, and its layer.conf only has one valid line:
BBPATH =. "${LAYERDIR}/../meta-poky:"

BTW, this layer was totally removed in 2017-09-14, but this is WRL9, so it is
still there.

The problem is after layerindex rebased to rebase-20181002 branch, there is an
additional checking in update.py:

   col = extract_value('BBFILE_COLLECTIONS', output)
   if not col:
       logger.error('Unable to find BBFILE_COLLECTIONS value in initial output')
       # Assume (perhaps naively) that it's an error specific to the layer
       continue

Which means layerindex would not add it to new layerbranch (RCPL0017) when
BBFILE_COLLECTIONS is not defined, there wasn't such a checking on layerindex'
old branch rebase-20170608, so meta-yocto was still added and released in
previous RCPLs.

We need go on releasing meta-yocto layer for RCPL9 since we released it in
previous RCPLs, so define a BBFILE_COLLECTIONS in meta-yoct/conf/layer.conf to
fix the problem.

(LOCAL REV: NOT UPSTREAM) -- WRLinux specific, upstream removed the layer totally.

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>