meta-yocto-bsp: Disable the glamor acceleration for the beaglebone
In the commit ec6d61cbddc3 ("mesa: enable native and nativesdk
variants"), the gallium is enabled for the mesa. Since we use the
modesetting driver for the Xorg on the beagleboard and the glamor
acceleration method is enabled by the modesetting driver, the gallium
DRI driver will be used after this commit. But the gallium DRI driver
only support 32bpp and we choose to use 16bpp on beaglebone in commit
e7434c17b4b3 ("meta-yocto-bsp: workaround the X malfunction on
beaglebone"), the mismatch between them causes the malfunction of the
Xorg. I have hacked the kernel to enable the 32bpp for the beagllebone,
but it has the following issues:
  1. The color is abnormal.
  2. The Xorg hang occasionally.
  3. The performance of using glamor acceleration based on gallium is
     pretty bad.
So I choose to disable the glamor on this board.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>